Fix ruff and mypy errors across Sprint 2-4 modules
Add N803, N815, S310 to ruff ignore (physics variable names, HTTP).
Fix E501 line length, E731 lambda, F401 unused imports, F841 unused
vars, SIM108 ternary, SIM300 yoda, RET504 return. Fix mypy: add
float() casts for np returns, Callable type for cost_fn, TYPE_CHECKING
import for UPDEEngine, tuple[float, ...] for weights.

+++ src/scpn_phase_orchestrator/coupling/spectral.py
@@ -96,4 +96,4 @@ def sync_convergence_rate(
         return 0.0
     lambda2 = fiedler_value(knm)
     k_eff = float(np.mean(knm[knm > 0])) if np.any(knm > 0) else 0.0
-    return k_eff * lambda2 * np.cos(gamma_max) / n
+    return float(k_eff * lambda2 * np.cos(gamma_max) / n)
